refactor search components and add a club search component
