--- src/components/CoffeeShopCard/CoffeeShopCard.jsx
@@ -1,9 +1,0 @@
-const CoffeeShopCard = () => {
-  return ( 
-    <div className="shop-card">
-      <h1>I am a shop card</h1>
-    </div>
-  )
-}
- 
-export default CoffeeShopCard
