+++ src/components/ShopSearchForm/ShopSearchForm.module.css
@@ -1,0 +1,47 @@
+.form-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  padding: 15px;
+  padding-bottom: 50px;
+  border: 3px solid white;
+  margin-top: 100px;
+  width: 90%;
+}
+
+.search-clubs {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+  /* gap: 20px; */
+  border: 3px solid white;
+  border-radius: 7px;
+  width: 500px;
+  height: 90px;
+  margin-top: 20px;
+  background-color: darkslategray;
+}
+.p-tag {
+  font-size: 20px;
+  width: 500px;
+  text-align: center;
+}
+.search > input {
+  width: 50%;
+  height: 40px;
+  border: 3px solid #787878;
+  border-radius: 5px;
+}
+.search > button {
+  height: 40px;
+  width: 30%;
+  background-color: brown;
+  border-radius: 5px;
+  border: none;
+  color: white;
+  font-size: 15px;
+  font-weight: bold;
+  border: 3px solid #b0b0b0;
+}
